pyproject.toml
[project]
name = "opengate-alarms"
version = "0.1.0"
description = "Retrieval of alarms from OpenGate IoT platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "opengate-data>=1.12.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "respx>=0.22.0",
    "textual>=8.0.0",
]

[project.scripts]
opengate-tui = "opengate_alarms.tui.app:run"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opengate_alarms"]
